DRYer code
