+++ di_website/publications/models.py
@@ -41,7 +41,7 @@ from taggit.models import Tag, TaggedItemBase
 
 from .mixins import (
     FlexibleContentMixin, UniquePageMixin, PageSearchMixin, LegacyPageSearchMixin, ParentPageSearchMixin,
-    PublishedDateMixin, UUIDMixin, ReportChildMixin)
+    PublishedDateMixin, UUIDMixin, ReportChildMixin, FilteredDatasetMixin)
 from .utils import (
     ContentPanel, PublishedDatePanel, WagtailImageField,
     UUIDPanel, get_first_child_of_type, get_ordered_children_of_type, get_downloads)
@@ -235,7 +235,7 @@ class PublicationIndexPage(HeroMixin, Page):
         verbose_name = 'Publication Index Page'
 
 
-class PublicationPage(HeroMixin, PublishedDateMixin, ParentPageSearchMixin, UUIDMixin, Page):
+class PublicationPage(HeroMixin, PublishedDateMixin, ParentPageSearchMixin, UUIDMixin, FilteredDatasetMixin,Page):
 
     class Meta:
         verbose_name = 'Publication Page'
@@ -356,16 +356,6 @@ class PublicationPage(HeroMixin, PublishedDateMixin, ParentPageSearchMixin, UUID
         children += list(self.appendices)
         return filter(None, children)
 
-    @cached_property
-    def filtered_datasets(self):
-        results = []
-        all_pub_datasets = self.publication_datasets.all()
-        for pub_dataset in all_pub_datasets:
-            if type(pub_dataset.dataset.specific).__name__ == "DatasetPage":
-                results.append(pub_dataset)
-        return results
-
-
     def save(self, *args, **kwargs):
         super(PublicationPage, self).save(*args, **kwargs)
 
@@ -375,7 +365,7 @@ class PublicationPage(HeroMixin, PublishedDateMixin, ParentPageSearchMixin, UUID
             Redirect(old_path=old_path, redirect_page=self).save()
 
 
-class PublicationSummaryPage(HeroMixin, ReportChildMixin, FlexibleContentMixin, PageSearchMixin, UniquePageMixin, UUIDMixin, Page):
+class PublicationSummaryPage(HeroMixin, ReportChildMixin, FlexibleContentMixin, PageSearchMixin, UniquePageMixin, UUIDMixin, FilteredDatasetMixin, Page):
 
     class Meta:
         verbose_name = 'Publication Summary'
@@ -456,17 +446,8 @@ class PublicationSummaryPage(HeroMixin, ReportChildMixin, FlexibleContentMixin, 
     def page_data_downloads(self):
         return self.data_downloads.all()
 
-    @cached_property
-    def filtered_datasets(self):
-        results = []
-        all_pub_datasets = self.publication_datasets.all()
-        for pub_dataset in all_pub_datasets:
-            if type(pub_dataset.dataset.specific).__name__ == "DatasetPage":
-                results.append(pub_dataset)
-        return results
 
-
-class PublicationChapterPage(HeroMixin, ReportChildMixin, FlexibleContentMixin, PageSearchMixin, UUIDMixin, Page):
+class PublicationChapterPage(HeroMixin, ReportChildMixin, FlexibleContentMixin, PageSearchMixin, UUIDMixin, FilteredDatasetMixin, Page):
 
     class Meta:
         verbose_name = 'Publication Chapter'
@@ -570,17 +551,8 @@ class PublicationChapterPage(HeroMixin, ReportChildMixin, FlexibleContentMixin, 
     def page_data_downloads(self):
         return self.data_downloads.all()
 
-    @cached_property
-    def filtered_datasets(self):
-        results = []
-        all_pub_datasets = self.publication_datasets.all()
-        for pub_dataset in all_pub_datasets:
-            if type(pub_dataset.dataset.specific).__name__ == "DatasetPage":
-                results.append(pub_dataset)
-        return results
 
-
-class PublicationAppendixPage(HeroMixin, ReportChildMixin, FlexibleContentMixin, PageSearchMixin, UUIDMixin, Page):
+class PublicationAppendixPage(HeroMixin, ReportChildMixin, FlexibleContentMixin, PageSearchMixin, UUIDMixin, FilteredDatasetMixin, Page):
 
     class Meta:
         verbose_name = 'Publication Appendix'
@@ -679,17 +651,8 @@ class PublicationAppendixPage(HeroMixin, ReportChildMixin, FlexibleContentMixin,
     def page_data_downloads(self):
         return self.data_downloads.all()
 
-    @cached_property
-    def filtered_datasets(self):
-        results = []
-        all_pub_datasets = self.publication_datasets.all()
-        for pub_dataset in all_pub_datasets:
-            if type(pub_dataset.dataset.specific).__name__ == "DatasetPage":
-                results.append(pub_dataset)
-        return results
 
-
-class LegacyPublicationPage(HeroMixin, PublishedDateMixin, LegacyPageSearchMixin, Page):
+class LegacyPublicationPage(HeroMixin, PublishedDateMixin, LegacyPageSearchMixin, FilteredDatasetMixin, Page):
 
     class Meta:
         verbose_name = 'Legacy Publication'
@@ -794,17 +757,8 @@ class LegacyPublicationPage(HeroMixin, PublishedDateMixin, LegacyPageSearchMixin
     def page_data_downloads(self):
         return self.data_downloads.all()
 
-    @cached_property
-    def filtered_datasets(self):
-        results = []
-        all_pub_datasets = self.publication_datasets.all()
-        for pub_dataset in all_pub_datasets:
-            if type(pub_dataset.dataset.specific).__name__ == "DatasetPage":
-                results.append(pub_dataset)
-        return results
 
-
-class ShortPublicationPage(HeroMixin, PublishedDateMixin, FlexibleContentMixin, PageSearchMixin, UUIDMixin, Page):
+class ShortPublicationPage(HeroMixin, PublishedDateMixin, FlexibleContentMixin, PageSearchMixin, UUIDMixin, FilteredDatasetMixin, Page):
 
     class Meta:
         verbose_name = 'Short Publication'
@@ -924,15 +878,6 @@ class ShortPublicationPage(HeroMixin, PublishedDateMixin, FlexibleContentMixin, 
             if block.block_type == 'section_heading':
                 sections.append(block)
         return sections
-
-    @cached_property
-    def filtered_datasets(self):
-        results = []
-        all_pub_datasets = self.publication_datasets.all()
-        for pub_dataset in all_pub_datasets:
-            if type(pub_dataset.dataset.specific).__name__ == "DatasetPage":
-                results.append(pub_dataset)
-        return results
 
 
 class PublicationPageRelatedLink(OtherPageMixin):
